feat: improve commit view and ui lib
